move suse specific salt additions to custom namespace
