+++ java/code/src/com/suse/manager/reactor/messaging/ImageDeployedEventMessage.java
@@ -16,7 +16,7 @@
 package com.suse.manager.reactor.messaging;
 
 import com.redhat.rhn.common.messaging.EventMessage;
-import com.suse.manager.webui.utils.salt.ImageDeployedEvent;
+import com.suse.manager.webui.utils.salt.custom.ImageDeployedEvent;
 import org.apache.commons.lang3.builder.ToStringBuilder;
 
 /**
