+++ java/code/src/com/suse/manager/reactor/messaging/ImageDeployedEventMessageAction.java
@@ -24,7 +24,7 @@ import com.redhat.rhn.domain.token.ActivationKey;
 import com.redhat.rhn.domain.token.ActivationKeyFactory;
 import com.suse.manager.reactor.utils.ValueMap;
 import com.suse.manager.webui.services.iface.SystemQuery;
-import com.suse.manager.webui.utils.salt.ImageDeployedEvent;
+import com.suse.manager.webui.utils.salt.custom.ImageDeployedEvent;
 import org.apache.log4j.Logger;
 
 import java.util.Optional;
