+++ java/code/src/com/suse/manager/reactor/messaging/RegisterMinionEventMessage.java
@@ -15,7 +15,7 @@
 package com.suse.manager.reactor.messaging;
 
 import com.redhat.rhn.common.messaging.EventMessage;
-import com.suse.manager.webui.utils.salt.MinionStartupGrains;
+import com.suse.manager.webui.utils.salt.custom.MinionStartupGrains;
 
 import java.util.Optional;
 /**
