+++ java/code/src/com/suse/manager/webui/services/impl/SaltSSHService.java
@@ -37,7 +37,7 @@ import com.suse.manager.webui.utils.SaltRoster;
 import com.suse.manager.webui.utils.SaltState;
 import com.suse.manager.webui.utils.SaltTop;
 import com.suse.manager.webui.utils.gson.BootstrapParameters;
-import com.suse.manager.webui.utils.salt.MgrActionChains;
+import com.suse.manager.webui.utils.salt.custom.MgrActionChains;
 import com.suse.manager.webui.utils.salt.State;
 import com.suse.salt.netapi.AuthModule;
 import com.suse.salt.netapi.calls.LocalCall;
